Reorganize TOC stuff
